+++ packages/client/src/types.ts
@@ -4,20 +4,25 @@ import type {
   VideoDimension,
 } from './gen/video/sfu/models/models';
 import type {
+  AudioSettingsRequestDefaultDeviceEnum,
   CallRecordingStartedEventRecordingTypeEnum,
   JoinCallRequest,
   MemberResponse,
   OwnCapability,
   ReactionResponse,
-  AudioSettingsRequestDefaultDeviceEnum,
   StartRecordingRequest,
   StartRecordingResponse,
 } from './gen/coordinator';
 import type { StreamClient } from './coordinator/connection/client';
+import type {
+  RejectReason,
+  StreamClientOptions,
+  TokenProvider,
+  User,
+} from './coordinator/connection/types';
 import type { Comparator } from './sorting';
 import type { StreamVideoWriteableStateStore } from './store';
 import { AxiosError } from 'axios';
-import { RejectReason } from './coordinator/connection/types';
 
 export type StreamReaction = Pick<
   ReactionResponse,
@@ -333,6 +338,28 @@ export type CallConstructor = {
    */
   clientStore: StreamVideoWriteableStateStore;
 };
+
+type StreamVideoClientBaseOptions = {
+  apiKey: string;
+  options?: StreamClientOptions;
+};
+
+type StreamVideoClientOptionsWithoutUser = StreamVideoClientBaseOptions & {
+  user?: undefined;
+  token?: never;
+  tokenProvider?: never;
+};
+
+type StreamVideoClientOptionsWithUser = StreamVideoClientBaseOptions & {
+  user: User;
+} & (
+    | { token: string; tokenProvider?: TokenProvider }
+    | { token?: string; tokenProvider: TokenProvider }
+  );
+
+export type StreamVideoClientOptions =
+  | StreamVideoClientOptionsWithoutUser
+  | StreamVideoClientOptionsWithUser;
 
 export type CallRecordingType = CallRecordingStartedEventRecordingTypeEnum;
 export type StartCallRecordingFnType = {